+++ changelog.md
@@ -4,3 +4,8 @@
 - prepare to use axios
 - add mock.js (available)
 - create api.js (not test)
+
+## [0.1.1] - 2019.7.6
+- axios api create
+- create warehousePreview.vue
+- create warehouseDetail.vue